add fixes after linters, README.md, add catalog/migrations/*(help_text in Product)
