+++ catalog/forms.py
@@ -2,32 +2,16 @@ import os
 import re
 from typing import Any, Union
 
-from django import forms
 from django.core.exceptions import ValidationError
 from django.core.files import File
 from django.core.files.uploadedfile import SimpleUploadedFile, UploadedFile
-from django.forms import BooleanField, ChoiceField, ModelForm
+from django.forms import ModelForm
 from django.template.defaultfilters import filesizeformat
 
 from config import settings
+from core.mixins import StyleFormMixin
 
 from .models import Category, Product
-
-
-class StyleFormMixin(forms.Form):
-    """Миксин для стилизации полей формы"""
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        """Инициализация миксина с настройкой атрибутов виджетов"""
-        super().__init__(*args, **kwargs)
-
-        for fild_name, fild in self.fields.items():
-            if isinstance(fild, BooleanField):
-                fild.widget.attrs["class"] = "form-check-input"
-            elif isinstance(fild, ChoiceField):
-                fild.widget.attrs["class"] = "form-select"
-            else:
-                fild.widget.attrs["class"] = "form-control"
 
 
 class ProductForm(StyleFormMixin, ModelForm):
@@ -37,7 +21,9 @@ class ProductForm(StyleFormMixin, ModelForm):
 
     class Meta:
         model = Product
-        exclude = ["owner", ]
+        exclude = [
+            "owner",
+        ]
 
     def clean_field(self, field_name: str) -> Any:
         """Общий метод валидации текстовых полей на запрещенные слова"""
@@ -107,7 +93,7 @@ class ProductModeratorForm(StyleFormMixin, ModelForm):
 
     class Meta:
         model = Product
-        fields = ("status", )
+        fields = ("status",)
 
 
 class CategoryForm(StyleFormMixin, ModelForm):
