+++ users/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
 
-from users.forms import CustomUserCreationForm
 from users.models import User
 
 
@@ -10,7 +9,7 @@ class UserAdmin(admin.ModelAdmin):
     пользователями, с возможностью фильтрации и поиска."""
 
     list_display = ("id", "email", "phone_number")
-    list_filter = ("email", )
+    list_filter = ("email",)
     search_fields = (
         "email",
         "phone_number",
