+++ users/management/commands/csu.py
@@ -5,10 +5,11 @@ from django.core.management.base import BaseCommand
 
 
 class Command(BaseCommand):
-    """ Команда для создания суперпользователя
-    с предопределенными учетными данными """
+    """Команда для создания суперпользователя
+    с предопределенными учетными данными"""
+
     def handle(self, *args: Any, **options: Any) -> None:
-        """ Добавляет суперпользователю email и пароль """
+        """Добавляет суперпользователю email и пароль"""
         User = get_user_model()
         user = User.objects.create(email="superuser@example.com")
         user.set_password("123qwer")
